requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-marshmallow==1.3.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
PyJWT==2.10.1
pymongo==4.13.0
python-decouple==3.8
Werkzeug==3.1.3
python-dotenv==1.0.1
flask-cors==4.0.0
flasgger==0.9.7.1
PyJWT==2.8.0
